backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
Django==2.2.7
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
